removed all HTML from dieroll.doRoll
